requirements.txt
Flask
Pillow
numpy
matplotlib
